+++ lib/oxaiworkers/tool/pixels.rb
@@ -19,7 +19,7 @@ module OxAiWorkers
           property :prompt, type: 'string', description: I18n.t('oxaiworkers.tool.pixels.generate_image.prompt'),
                             required: true
           if worker.sizes.length > 1
-            property :size, type: 'string', description: I18n.t('oxaiworkers.tool.pixels.generate_image.size'),
+            property :size, type: %w[string null], description: I18n.t('oxaiworkers.tool.pixels.generate_image.size'),
                             enum: worker.sizes
           end
           if current_dir.present?
@@ -28,7 +28,7 @@ module OxAiWorkers
                                  required: true
           end
           if worker.qualities.length > 1
-            property :quality, type: 'string', description: I18n.t('oxaiworkers.tool.pixels.generate_image.quality'),
+            property :quality, type: %w[string null], description: I18n.t('oxaiworkers.tool.pixels.generate_image.quality'),
                                enum: worker.qualities
           end
         end