+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.1] - 2025-05-13
 
 - Property `required` parameter is now `true` by default in `define_function`
+- Fixed `type` parameter for `define_function`
 
 ## [1.0.0] - 2025-05-12